refactor(governance): extract vote card content
